refactor: extract shared helpers and replace once_cell with std LazyLock
- Extract duplicated batch/record helper functions (batch_records_to_py,
  record_to_meta, batch_record_meta) from client.rs and async_client.rs
  into a shared record_helpers.rs module
- Replace once_cell::sync::Lazy with std::sync::LazyLock (stable since
  Rust 1.80) and remove once_cell direct dependency
- Deduplicate TTL/meta computation in client.rs exists() and
  operate_ordered() methods
