+++ rust/Cargo.toml
@@ -14,4 +14,3 @@ aerospike = { git = "https://github.com/aerospike/aerospike-client-rust.git", ta
 aerospike-core = { git = "https://github.com/aerospike/aerospike-client-rust.git", tag = "v2.0.0-alpha.9" }
 tokio = { version = "1", features = ["full"] }
 futures = "0.3"
-once_cell = "1"
